Put CommonOptions.installTiller back in the right place.
